Moving websockets to another file, changed UI for game creation form:
