+++ backend/app.js
@@ -6,7 +6,6 @@ const poker = require('./routes/poker')
 const sudoku = require("./routes/sudoku");
 
 const app = express();
-
 app.use(cors({
   origin: "*",
 }));
@@ -16,9 +15,12 @@ app.use(express.urlencoded({
   extended: true
 }));
 
+
+
 app.use('/auth', auth);
 app.use('/poker', poker);
 app.use("/sudoku", sudoku);
+
 
 app.use((req, res, next) => {
     console.log("called2");
