+++ backend/routes/sudoku.js
@@ -1,10 +1,5 @@
 var express = require('express')
-var connection = require("../connection")
-var crypto = require("crypto");
-var utils = require("../utils");
-var Table = require("../modules/table");
 
 var router = express.Router()
-
 
 module.exports = router
