+++ frontend/src/components/sudoku/Board.js
@@ -4,7 +4,6 @@ import { io } from "socket.io-client";
 
 import Chatbox from "../common/Chatbox";
 import ControlButton from "../common/ControlButton";
-import { soundContext } from "../../contexts/soundContext";
 
 import "../../styles/board.scss"
 import "../../styles/chatbox.scss"
@@ -22,7 +21,7 @@ class Cell {
 export default function Board(props) {
     const [board, setBoard] = useState(createBoard(props.board));
     const [startingBoard, setStartingBoard] = useState(null);
-    const [baseIndex, setBaseIndex] = useState(getBase(props.board));
+    const [baseIndex, setBaseIndex] = useState(props.base);
     const [highlightIndex, setHighlightIndex] = useState(82);
     const [annotate, setAnnotate] = useState(false);
 
@@ -44,135 +43,12 @@ export default function Board(props) {
 
     const ALLOWED_INPUTS = [0,1,2,3,4,5,6,7,8,9];
 
-    let sound = useContext(soundContext)
-    
-    function useQuery() {
-        return new URLSearchParams(useLocation().search);
-    }
-
-    let query = useQuery();
-
-    useEffect(() => {
-        console.log("useeffect");
-        let socket = io(`${process.env.REACT_APP_API_URL}/sudoku`, {
-            auth: {
-                token: localStorage.getItem("token")
-            }
-        });
-
-        setSocket(socket);
-
-        socket.on("connect", () => {
-            console.log("connected");
-        })
-
-        if (query.get("create")) {
-            let difficulty = query.get("difficulty")
-            socket.emit("create", difficulty)
-        } else {
-            console.log("joining => ", room_code)
-            socket.emit("join", room_code);
-        }
-
-        socket.on("created", (code) => {
-            setRoomCode(code)
-            setWaiting(true)
-            setCreated(true)
-            console.log(code, room_code);
-        });
-
-        socket.on("joined", (userInformation) => {
-            setOpponent(userInformation);
-        });
-
-        socket.on("start", (data) => {
-            console.log("started", data, data.board);
-            setBaseIndex(getBase(data.board))
-            setBoard(createBoard(data.board))
-            setWaiting(false);
-            setStartingBoard(data.board);
-            sound("GameStarted");
-        });
-
-        socket.on("state", (board, opponentInfo) => {
-            console.log(opponentInfo);
-            setBaseIndex(getBase(board))
-            setBoard(createBoard(board));
-            setOpponent(opponentInfo.user);
-            setOpponentScore(opponentInfo.score);
-            setWaiting(false);
-            setStartingBoard(board);
-            sound("GameStarted");
-        });
-
-        //this is probably horrible, just figure out a way to unmount and have it join normally
-        //i cant do this anymore
-        //delete it ALL
-        socket.on("redirect", (new_code, data) => {
-            console.log("REDIRECTING => ", new_code)
-            setBaseIndex(getBase(data.board))
-            setBoard(createBoard(data.board))
-            setWaiting(false);
-            setStartingBoard(data.board);
-            setRoomCode(new_code);
-            setRematch(new_code);
-        });
-
-        socket.on("Filled square", (operation) => {
-            console.log("???")
-            if (operation == "add") {
-                setOpponentScore(state => state + 1);
-
-            } else if (operation == "subtract") {
-                setOpponentScore(state => state - 1);
-
-            }
-            
-        })
-
-        socket.on("Winner", (result) => {
-            setResult(result);
-        })
-
-        socket.on("chat", (message) => {
-            console.log("CHAT => ", message)
-            setMessages(prevMessages => [...prevMessages, message]);
-        });
-
-        socket.on("err", (message) => {
-            console.log("ERR: ", message);
-        });
-
-        return () => {
-            console.log("disconnecting")
-            socket.disconnect();
-        }
-    }, [rematch])
-
-    
-
     useEffect(() => {
         window.addEventListener("keydown", handleEventInput);
         return () => {
             window.removeEventListener("keydown", handleEventInput);
         };
     }, [highlightIndex, annotate]);
-
-    function handleRematch() {
-        socket.emit("rematch", !rematchStatus);
-        setRematchStatus(!rematchStatus);
-    }
-
-    function handleChatInput(e) {
-        setHighlightIndex(82);
-        setChatInput(e.target.value);
-    }
-
-    function handleChatSubmit(e) {
-        e.preventDefault();
-        socket.emit("chat", chatInput);
-        setChatInput("");
-    }
 
     function handleEventInput({key}) {
         if (
@@ -247,42 +123,6 @@ export default function Board(props) {
 
     function resetBoard() {
         setBoard(createBoard(startingBoard))
-    }
-
-    function getBaseJson(board_json) {
-
-        let indexs = [];
-
-        board_json.forEach((square, index) => {
-            if (square.value != 0) {
-                console.log(index, typeof index)
-                indexs.push(index.toString());
-            }
-        })
-
-        return indexs;
-    }
-
-    function getBase(board) {
-
-        if (typeof board == "string") {
-            let indexs = [];
-
-            for (let i = 0; i < board.length; i++) {
-                if (board[i] != 0) {
-                    indexs.push(i.toString());
-                }
-            }
-    
-            return indexs;
-    
-        } else if (typeof board == "object") {
-            return getBaseJson(board);
-        } else {
-            return getBase(BOARD_DEFAULT_INDEX);
-        }
-
-
     }
 
     function handleCellClick(e) {
@@ -423,38 +263,8 @@ export default function Board(props) {
     } else {
         total = 0;
     }
-    
-
-    if (result == null) {
-        scoreText = <span className="opponent-score">Opponent: {opponentScore}/{total}</span> 
-        
-    } else if (result == "win") {
-        scoreText = 
-        <div style={{display:"flex", flexDirection: "column"}}>
-            <span className="opponent-score win">Opponent: {opponentScore}/81</span> 
-            <span className="opponent-score win">You Won!</span> 
-            <ControlButton 
-                handleClick={() => handleRematch()}
-                name={"Rematch"}
-                class_on={"rematch-button active"}
-                class_off={"rematch-button inactive"}
-            />
-        </div>
-    } else {
-        scoreText = <div>            
-                        <span className="opponent-score lose">Opponent: Won</span> 
-                        <ControlButton 
-                            handleClick={() => handleRematch()}
-                            name={"Rematch"}
-                            class_on={"rematch-button active"}
-                            class_off={"rematch-button inactive"}
-                        />
-                    </div>
-        
-    }
 
     return (
-        <div className="board-top-container">
             <div className="center-board-container">
                 {created ? <Redirect to={`/sudoku/${room_code}`} /> : null }
                 {rematch ? <Redirect to={`/sudoku/${rematch}`} /> : null}
@@ -488,31 +298,5 @@ export default function Board(props) {
                 </div>
                 
             </div>
-            <div className="right-chat-container">
-                <div className="chat-top-container">
-                    <div class="opponent-container">
-                        <div class="user-info">
-                            <div class="user">
-                                <span class="name">{opponent == null ? "Guest" : opponent.username}</span>
-                            </div>
-                        </div>
-                    </div>
-                    <div class="chat-container">
-                        <Chatbox messages={messages} />
-                        <form onSubmit={(e) => handleChatSubmit(e)}>
-                            <input class="chat-input" 
-                                onSubmit={(e) => handleChatSubmit(e)} 
-                                placeholder="Start typing here..." value={chatInput} 
-                                onChange={(e) => handleChatInput(e)}
-                                onClick={() => setHighlightIndex(82)}>
-                            </input>
-                            <input type="submit" style={{display: "none"}}></input>
-                        </form>
-
-                    </div>
-                </div>
-                
-            </div>
-        </div>
     )
 }
