+++ frontend/src/styles/blockpage.scss
@@ -9,5 +9,5 @@
     top: 0;
     left: 0;
     right: 0;
-    z-index: 4;
+    z-index: 1;
 }
