+++ frontend/src/styles/login.scss
@@ -12,6 +12,8 @@
     padding: 20px;
     align-items: center;
   }
+
+  
   
   .login-form {
     display: flex;
