+++ frontend/src/styles/slider.scss
@@ -1,6 +1,6 @@
 input[type=range] {
     -webkit-appearance: none;
-    width: 80%;
+    width: 100%;
     background: transparent;
 }
